wifi: ath11k: handle thermal device registeration together with MAC
We are see below warning when resume,
[12841.841607] hwmon hwmon11: PM: parent phy0 should not be sleeping

This is because ath11k_thermal_register() is called in
ath11k_pci_pm_resume(), while at that time ar->hw->wiphy->dev
is still sleeping.

To fix it, change to handle thermal device together with MAC
register/unregister which are not involved in suspend/resume.

Tested-on: WCN6855 hw2.1 PCI WLAN.HSP.1.1-03125-QCAHSPSWPL_V1_V2_SILICONZ_LITE-3.6510.23

Signed-off-by: Baochen Qiang <quic_bqiang@quicinc.com>
Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>