wifi: ath12k: Fix the QoS control field offset to build QoS header
Currently, in the mac80211 layer, received EAPOL packets are dropped
when the HT control field is present in the QoS header. This occurs
because an incorrect QoS control field offset is used to build
the Qos header in the MSDU data, leading to a corrupted header in the
mac80211 layer. To fix this, use ieee80211_get_qos_ctl() to get the
correct QoS control offset from the MSDU data.

Signed-off-by: Ramasamy Kaliappan <quic_rkaliapp@quicinc.com>
Signed-off-by: Nithyanantham Paramasivam <quic_nithp@quicinc.com>
Signed-off-by: Raj Kumar Bhagat <quic_rajkbhag@quicinc.com>
Signed-off-by: Jeff Johnson <jeff.johnson@oss.qualcomm.com>