wifi: cfg80211: check radio iface combination for multi radio per wiphy
Currently, wiphy_verify_combinations() fails for the multi-radio per wiphy.
This is because the global interface combination encompasses the
capabilities of all radio combinations. For multi-radio per wiphy,
interface combination verification needs to be performed for radio specific
interface combinations. This is necessary because the global interface
combination aggregate the capabilities of all radio combinations.

Tested-on: QCN9274 hw2.0 PCI WLAN.WBE.1.0.1-00029-QCAHKSWPL_SILICONZ-1

Signed-off-by: Karthikeyan Periyasamy <quic_periyasa@quicinc.com>
Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>