wifi: ath12k: Add helpers for multi link peer creation and deletion
Add helper apis for multi link peer addition, deletion.
Add address validation to ensure we are not creating link
peers (belonging to different clients) with same MLD address.
To aid in this validation for faster lookup, add a new list
of ML peers at ah level and use the same for parsing for
the above address validation use cases.

Signed-off-by: Sriram R <quic_srirrama@quicinc.com>
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>
Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>