wifi: ath12k: do not return invalid link id for scan link
When a scan request is received, driver selects a link id
for which the arvif can be mapped. Same link is also
used for getting the link conf address. Currently, we return
0 as link id for a non ML vif, which is correct since that is
the default link id.
Also when any of the link vif is active and the scan request
is for a channel in the active link we return its link id.
But, when we don't hit both of the above cases(i.e not a ML
vif or no active link vif for the channel is present) we
currently return 0 as the link id.
This might not work out always, eg., when only one
link(ex. linkid = 1) is added to vif, then we won't find any
link conf for link id 0 in the vif resulting in scan failure.

During AP bringup, such scan failure causes bringup issues.
Hence avoid sending link id 0 as default. Rather use a default
link for scan and default link address for the same. This scan
vdev will either be deleted if another scan is requested on same
vif or when AP is broughtup on same link or during interface cleanup.

Signed-off-by: Sriram R <quic_srirrama@quicinc.com>
Signed-off-by: Rameshkumar Sundaram <quic_ramess@quicinc.com>
Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>