wifi: ath12k: symmetrize scan vdev creation and deletion during HW scan
Currently, the hardware scan is initiated in the driver on scan link (15).
After mapping to the appropriate radio based on the scan frequency, the
vdev is created and the scan begins. However, the vdev is only deleted when
channel assignment is about to occur after the scan. Additionally, it is
also deleted if a new scan is requested on the same interface but the
underlying radio differs in the new request.

This imbalance leads to various hardware scan issues, especially when a
non-MLO and MLO combination exists. In such cases, the latter tries to
skip the scan and proceed with channel assignment while the former is still
scanning, causing a firmware assert.

To address this issue, symmetrize the scan vdev creation and deletion
during hardware scan operations. This means creating a vdev when the scan
starts and deleting it once the scan is completed or aborted.

While at this, add a few debug prints in scan handling and a few empty
lines for better code read.

Patch: [New patch]

Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>