wifi: ath12k: add ATH12K_FW_FEATURE_MLO capability firmware feature
To maintain backward compatibility with older firmware versions, introduce
a new feature bit, ATH12K_FW_FEATURE_MLO, to identify whether the firmware
supports MLO.

If the firmware-X.bin does not have this bit advertised in the feature,
then MLO capability in the host will be disabled.

This applies only for QCNXXXX chipsets.

Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>