wifi: ath12k: add can_activate_links mac ops
When operating as a ML station, mac80211 initially activates only one link
in the driver until the peer is authorized. Once the state changes to 4
(authorized), the driver should call mac80211 API to activate all other
partner links.

Before doing so, mac80211 checks if the driver supports activating links
via the can_activate_links mac80211_ops.

Therefore, add support for this mac80211_ops and call the API to activate
the links once the state changes to 4.

Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>