wifi: ath12k: Parse multiple device information from device tree
Currently, a single device is part of the device group abstraction.
However, for multi-link operations, multiple devices need to be combined.
This multi-device grouping is done via WSI (WLAN Serial Interface), which
is described in the device tree. Information about different WSI groups
and the number of devices involved in each group can be parsed from the
device tree.

Add changes to parse the device tree and determine WSI information, such
as the different WSI groups and the number of devices per WSI group.
Assign WSI index zero to the WSI controller device (to synchronize the
clock among the devices within the WSI group), and increment the WSI index
of each device in the order of the WSI connection.

Co-developed-by: Harshitha Prem <quic_hprem@quicinc.com>
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>
Signed-off-by: Raj Kumar Bhagat <quic_rajkbhag@quicinc.com>
Signed-off-by: Kalle Valo <quic_kvalo@quicinc.com>