wifi: ath12k: select scan device based on ar frequency range for splitphy
Currently while selecting the radio(ar) for initiating HW scan the first
radio which has the scan requested channel's band will be selected.
In case of splitphy cards,a single band(say 5GHz) is split across two
radios (5GHz low channel 36-64 and 5GHz high channel 100-177) and above
selection will result in 5GHz low radio (ar) being selected when a scan is
requested on a 5GHz channel. This scan request will eventually fail as HW
will not able to scan on given frequency which it is not capable on that
radio.

Fix this by selecting the scan radio based on the frequency range of each
radio.
While at it add a debug message to print the frequency range of each radio
during mac registration

Signed-off-by: Rameshkumar Sundaram <rameshkumar.sundaram@oss.qualcomm.com>
Signed-off-by: Raj Kumar Bhagat <quic_rajkbhag@quicinc.com>
Signed-off-by: Jeff Johnson <jeff.johnson@oss.qualcomm.com>