wifi: ath12k: Split scan request for split band device
When supplicant initiates scan req for split band device,
mac80211 generates an single scan req to driver with all the
channel frequencies. This is because mac80211 categorize the
the frequencies based on band only.

Driver checks the first freq from the channel list and initiates
scan to firmware with all the frequencies. Firmware rejects the freq
which are not supported. As a result of this, Supplicant receives
the scan results from only one device and associates on one link.

Fix this by splitting the scan request in driver and perform
multiple scans, send completion notification to mac80211 after
split scans are done.

Signed-off-by: Vignesh C <quic_vignc@quicinc.com>
Signed-off-by: Rameshkumar Sundaram <rameshkumar.sundaram@oss.qualcomm.com>
Signed-off-by: Raj Kumar Bhagat <quic_rajkbhag@quicinc.com>
Signed-off-by: Jeff Johnson <jeff.johnson@oss.qualcomm.com>