wifi: ath12k: DVLAN multicast to be treated as unicast in Tx op
Since DVLAN traffic is software encrypted and encapsulated, mac80211
handles the duplication on each link for MLO DVLAN multicast.

Currently in mac Tx op, all multicast frames are duplicated and are
assigned a mcbc_gsn. This is causing DVLAN multicast traffic to be
dropped in HW.

Treat DVLAN multicast traffic as unicast since the duplication on each
link is handled by mac80211.

Signed-off-by: Muna Sinada <muna.sinada@oss.qualcomm.com>
Signed-off-by: Jeff Johnson <jeff.johnson@oss.qualcomm.com>