wifi: ath12k: disable pdev for non supported country
Following WARN_ON is thrown for LK/CN country,

[ 1574.376739] Call trace:
[ 1574.382042]  ath12k_reg_update_chan_list+0xac/0x2c8 [ath12k]
[ 1574.384215]  ath12k_mac_drain_tx+0x4d0/0x690 [ath12k]
[ 1574.390122]  drv_start+0x44/0x54 [mac80211]
[ 1574.395066]  ieee80211_do_open+0x154/0x860 [mac80211]
[ 1574.399060]  ieee80211_do_open+0x838/0x860 [mac80211]
[ 1574.404261]  __dev_open+0xe4/0x164
[ 1574.409291]  __dev_change_flags+0xec/0x19c
[ 1574.412590]  dev_change_flags+0x20/0x5c
[ 1574.416671]  devinet_ioctl+0x258/0x580
[ 1574.420402]  inet_ioctl+0x678/0x6ac
[ 1574.424223]  sock_do_ioctl+0x4c/0x61c
[ 1574.427607]  sock_ioctl+0x80c/0x870
[ 1574.431427]  vfs_ioctl+0x24/0x48
[ 1574.434724]  do_vfs_ioctl+0xa3c/0xaa8
[ 1574.438197]  ksys_ioctl+0x44/0x74
[ 1574.441756]  __arm64_sys_ioctl+0x18/0x2c8
[ 1574.445057]  el0_svc_common.constprop.0+0x98/0x114
[ 1574.449048]  el0_svc_handler+0x18/0x20
[ 1574.453735]  el0_svc+0x8/0x500
[ 1574.457466] ---[ end trace 90f7e1ac33aba033 ]---

In MLO configuration, ath12k_mac_radio_start iterates through all the
radio’s and makes the ar state 'ON'. Eventhough some bands are not
supported in certain countries, ath12k_reg_update_chan_list tries to
update
the channel list for all the active pdev's and ends up in the warn_on
for non-supported band.

To fix this, disabling the pdev when the num of channels
in a band for a particular country is zero.

Signed-off-by: Karthik M <quic_karm@quicinc.com>
Signed-off-by: Manish Dharanenthiran <manish.dharanenthiran@oss.qualcomm.com>
Signed-off-by: Raj Kumar Bhagat <quic_rajkbhag@quicinc.com>
Signed-off-by: Jeff Johnson <jeff.johnson@oss.qualcomm.com>