wifi: ath12k: ath12k: Fix get peer from peer id instead of ast index
Currently in function ath12k_dp_mon_rx_update_user_stats, we are
fetching peer by ast index look up. But in monitor mode the ast
index is maped to 0 to 15 bits and in latest HTT_T2H_MSG_TYPE_PEER_MAP3
ast hash is mapped to 16 to 19 bits.
Due to this when multiple clients are connected we are not able to fetch
peer from ast index look up.

Hence, use peer_id to access peer instead of ast index.

Tested-on: QCN9274 hw2.0 PCI WLAN.WBE.1.3.1-00173-QCAHKSWPL_SILICONZ-1

Signed-off-by: Rajat Soni <quic_rajson@quicinc.com>
Signed-off-by: Maharaja Kennadyrajan <maharaja.kennadyrajan@oss.qualcomm.com>
Signed-off-by: Raj Kumar Bhagat <quic_rajkbhag@quicinc.com>
Signed-off-by: Jeff Johnson <jeff.johnson@oss.qualcomm.com>