wifi: mac80211: Allow scan on a link while operating in DFS on another link
Currently, if a link is operating on a DFS channel, -EBUSY is returned when
a new scan is requested. However, for Multi-Link Operation (MLO), multiple
hardware can be grouped under a single wiphy. In such cases, if a link is
operating on a DFS channel on one underlying hardware and a scan is
requested on a different underlying hardware, the scan can be allowed
simultaneously.

Add logic to check the underlying hardware used for the requested scan. If
the hardware on which DFS is already running is not being used, the scan
operation is allowed; otherwise, -EBUSY is returned.

For non-MLO scenarios, there is no change.

Signed-off-by: Raj Kumar Bhagat <quic_rajkbhag@quicinc.com>
Signed-off-by: Jeff Johnson <jeff.johnson@oss.qualcomm.com>