wifi: ath12k: correctly handle mcast packets for clients
Currently is_mcbc bit is set for packest send from client as
destination address (DA) is multicast/broadcast address, but packets
are actually unicast as receiver address (RA) is not multicast address.
Due to this packets are getting dropped.

Hence, reset the is_mcbc bit if interface type is AP.

Signed-off-by: Sarika Sharma <quic_sarishar@quicinc.com>
Signed-off-by: Raj Kumar Bhagat <quic_rajkbhag@quicinc.com>
Signed-off-by: Jeff Johnson <jeff.johnson@oss.qualcomm.com>