wifi: ath12k: Make compact tlv ops as the default ops for qcn
Make compact tlv ops as the default ops by registering them in
hw_params by default for qcn, and remove non-compact tlv ops and the
corresponding hal APIs qcn, as they won't be in use for qcn.

No changes for wcn hal ops as these changes are specific to qcn.

Tested-on: QCN9274 hw2.0 PCI WLAN.WBE.1.4.1-00199-QCAHKSWPL_SILICONZ-1
Tested-on: WCN7850 hw2.0 PCI WLAN.HMT.1.0.c5-00481-QCAHMTSWPL_V1.0_V2.0_SILICONZ-3

Signed-off-by: Pavankumar Nandeshwar <quic_pnandesh@quicinc.com>
Signed-off-by: Ripan Deuri <quic_rdeuri@quicinc.com>
Signed-off-by: Jeff Johnson <jeff.johnson@oss.qualcomm.com>