wifi: ath12k: Convert ath12k_dp member in ath12k_base to pointer
With Unified Driver framework, the Data Path (DP) specific device object
(ath12k_dp) need to support different hardware architectures. Keeping DP
structure as a pointer makes it dynamic and thus the allocation and
initialization is done as per the underlying hardware architecture.

In this change, convert ath12k_dp member in ath12k_base to a pointer,
enabling hardware-dependent allocations. This flexibility is essential for
future enhancements and hardware-specific initializations.

This patch (and the forthcoming patches) are intended to serve the purpose
of refactoring DP objects for the Unified Driver framework.

Tested-on: QCN9274 hw2.0 PCI WLAN.WBE.1.4.1-00199-QCAHKSWPL_SILICONZ-1
Tested-on: WCN7850 hw2.0 PCI WLAN.HMT.1.0.c5-00481-QCAHMTSWPL_V1.0_V2.0_SILICONZ-3

Signed-off-by: Ripan Deuri <quic_rdeuri@quicinc.com>
Signed-off-by: Jeff Johnson <jeff.johnson@oss.qualcomm.com>