wifi: ath12k: Attach and detach ath12k_dp_link_peer to ath12k_dp_peer
This patch is intended to define APIs to attach and detach ath12k_dp_link
peer to ath12k_dp_peer as and when it is created and deleted.

Background:

As part of peer modularization in Unified Driver Framework, station view
for the data path is shown below:

    ath12k_dp_peer
           |
           |-> ath12k_dp_link_peer
           |
           |-> ath12k_dp_link_peer
           |
           |-> ath12k_dp_link_peer

This ath12k_dp_peer contains the fields used in the per pkt Tx Rx paths
applicable across all the links and maintains an array of
ath12k_dp_link_peer.

Tested-on: QCN9274 hw2.0 PCI WLAN.WBE.1.4.1-00199-QCAHKSWPL_SILICONZ-1
Tested-on: WCN7850 hw2.0 PCI WLAN.HMT.1.0.c5-00481-QCAHMTSWPL_V1.0_V2.0_SILICONZ-3

Signed-off-by: Harsh Kumar Bijlani <quic_hbijlani@quicinc.com>
Signed-off-by: Jeff Johnson <jeff.johnson@oss.qualcomm.com>